refactor: add classification and rename signadyne->keysight
Rename driver and classes due to acquisition of Signadyne by Keysight and add classification (M3201A)--inherits-->(SD_Module)
